pyproject.toml
[tool.poetry]
name = "goode_bot_twitch"
version = "0.1.0"
description = ""
authors = ["cheeseborgers <goodcheeseburgers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
twitchio = "^2.3.0"
python-dotenv = "^0.20.0"
dateutils = "^0.6.12"
aiosqlite = "^0.17.0"
aiofiles = "^0.8.0"


[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
